refactor: move the window client's TypeScript helpers into modules
Phase 1 of issue #79, step 1 of 3. Moves the ~1,000 lines of exported
TypeScript helpers out of src/window-client.ts into eight cohesive modules
under src/window-client/ (drawing, live-ground, live-scatter,
live-visibility, live-camera, live-replay, rows, directory), each copied
verbatim with no renaming, reformatting, or reordering. src/window-client.ts
imports from the new modules and re-exports every name under its original
path, so every existing consumer (tests, e2e specs, src/window.ts) needs no
change. windowLivePointFootprints gains `export` (it is stringified by the
still-inline WINDOW_JS template) but is not re-exported from the facade — it
was never public before this change.

scripts/check-refusal-census.ts now skips the new src/window-client/
subtree the same way it already skipped window-client.ts: non-boundary
presentation code, not a caller-facing door.

The WINDOW_JS template literal is untouched and stays inline in this commit
(it moves to ordered program parts in the next commit), so it remains
byte-identical to main: sha256
8388f3b9b32ccb8b0c4c5c97ca9bd40a9d24e895c2abaaebbc69d0a60845f178,
548,524 bytes / 548,218 chars.

npm run typecheck, npm test (1901/1905 pass; the 4 failures are the known
Windows-only identity-client failures, issue #183), and npm run door:check
all pass.
