+++ scripts/check-refusal-census.ts
@@ -236,6 +236,12 @@ function sourceFiles(projectRoot: URL): string[] {
     .filter(entry => entry.endsWith('.ts'))
     .filter(entry => !IDENTITY_MODULES.has(basename(entry)))
     .filter(entry => !NON_BOUNDARY_MODULES.has(basename(entry)))
+    // Phase 1 of issue #79 split window-client.ts's non-boundary presentation
+    // code into src/window-client/**.ts (helper modules and ordered client-JS
+    // program parts). Every file under that directory is the same
+    // non-boundary surface window-client.ts itself is excluded above for,
+    // not a caller-facing door -- so the whole subtree is skipped the same way.
+    .filter(entry => !entry.startsWith('window-client/'))
     .map(entry => resolve(src, entry))
     .sort()
 }
